+++ crates/api_models/src/user_role.rs
@@ -133,8 +133,3 @@ pub struct AcceptInvitationRequest {
 pub struct DeleteUserRoleRequest {
     pub email: pii::Email,
 }
-
-#[derive(Debug, serde::Deserialize, serde::Serialize)]
-pub struct TransferOrgOwnershipRequest {
-    pub email: pii::Email,
-}