+++ crates/common_utils/src/events.rs
@@ -123,7 +123,7 @@ impl_api_event_type!(
         String,
         id_type::MerchantId,
         (id_type::MerchantId, String),
-        (&id_type::MerchantId, String),
+        (id_type::MerchantId, &String),
         (&id_type::MerchantId, &String),
         (&String, &String),
         (Option<i64>, Option<i64>, String),