+++ crates/router/src/connector/stripe.rs
@@ -157,6 +157,8 @@ impl ConnectorValidation for Stripe {
             PaymentMethodDataType::ApplePay,
             PaymentMethodDataType::GooglePay,
             PaymentMethodDataType::AchBankDebit,
+            PaymentMethodDataType::BacsBankDebit,
+            PaymentMethodDataType::BecsBankDebit,
             PaymentMethodDataType::SepaBankDebit,
             PaymentMethodDataType::Sofort,
             PaymentMethodDataType::Ideal,