+++ crates/router/src/core/fraud_check/flows/transaction_flow.rs
@@ -29,6 +29,23 @@ impl
         FraudCheckResponseData,
     > for FrmData
 {
+    #[cfg(all(feature = "v2", feature = "customer_v2"))]
+    async fn construct_router_data<'a>(
+        &self,
+        _state: &SessionState,
+        _connector_id: &str,
+        _merchant_account: &domain::MerchantAccount,
+        _key_store: &domain::MerchantKeyStore,
+        _customer: &Option<domain::Customer>,
+        _merchant_connector_account: &helpers::MerchantConnectorAccountType,
+        _merchant_recipient_data: Option<MerchantRecipientData>,
+    ) -> RouterResult<
+        RouterData<frm_api::Transaction, FraudCheckTransactionData, FraudCheckResponseData>,
+    > {
+        todo!()
+    }
+
+    #[cfg(all(any(feature = "v1", feature = "v2"), not(feature = "customer_v2")))]
     async fn construct_router_data<'a>(
         &self,
         _state: &SessionState,
@@ -50,9 +67,7 @@ impl
                 id: "ConnectorAuthType".to_string(),
             })?;
 
-        let customer_id = customer
-            .to_owned()
-            .map(|customer| customer.get_customer_id());
+        let customer_id = customer.to_owned().map(|customer| customer.customer_id);
 
         let payment_method = self.payment_attempt.payment_method;
         let currency = self.payment_attempt.currency;