+++ crates/router/src/core/mandate.rs
@@ -283,7 +283,7 @@ pub async fn mandate_procedure<F, FData>(
     resp: &types::RouterData<F, FData, types::PaymentsResponseData>,
     customer_id: &Option<id_type::CustomerId>,
     pm_id: Option<String>,
-    merchant_connector_id: Option<String>,
+    merchant_connector_id: Option<id_type::MerchantConnectorAccountId>,
     storage_scheme: MerchantStorageScheme,
     payment_id: &id_type::PaymentId,
 ) -> errors::RouterResult<Option<String>>